+++ changelog.md
@@ -12,6 +12,7 @@
 - Optimized option to hide the welcome notice
 - Optimized security for handling of local files in hosting
 - Prevent usage of credentials for URLs in the import dialog if the website does not use SSL
+- Extended the dialog after saving an external source with more options
 - Updated support for Abilities API
 - Updated dependencies